Add a version to bundles in the database
NOTE: This is a *database-only* change; no other code uses this yet. The
defaults ensure that this will be transparent to other parts of the
system.

Versions are semver-style "major.minor.patch" strings, transformed via a
custom `Ecto.Type` implementation into a 3-element integer array in the
database (this makes it easy to sort, for instance).